ETF-Porfolio-Builder/requirements.txt
# Core data analysis and financial libraries
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Financial data and APIs
yfinance>=0.2.18
alpha_vantage>=2.3.1
pandas-datareader>=0.10.0
financials-extension>=0.1.0

# Portfolio optimization
cvxpy>=1.3.0
pypfopt>=1.5.0
quantlib>=1.32

# Statistical modeling and Monte Carlo
scikit-learn>=1.3.0
statsmodels>=0.14.0
arch>=6.2.0

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0
dash>=2.12.0
dash-bootstrap-components>=1.4.0

# Web scraping and APIs
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.11.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=8.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
